restructure wrappers into current and old. begin testing and removing HORTON2
