+++ chemtools/wrappers/grid.py
@@ -21,79 +21,36 @@
 #
 # --
 """Grid Wrapper Module."""
-
-
-import numpy as np
-
-from horton import BeckeMolGrid
+from grid.becke import BeckeWeights
+from grid.molgrid import MolGrid
+from grid.onedgrid import HortonLinear
 from chemtools.wrappers.molecule import Molecule
-
 
 __all__ = ['MolecularGrid']
 
-
-class MolecularGrid(object):
+#TODO: How to implement compute spherical average functions?
+class MolecularGrid:
     """Becke-Lebedev molecular grid for numerical integrations."""
 
-    def __init__(self, coordinates, numbers, pseudo_numbers, specs='medium', k=3, rotate=False):
-        """Initialize class.
+    def __init__(self, coordinates, numbers, pseudo_numbers,
+                 specs='medium', k=3, points_of_angular=110, rotate=False):
 
-        Parameters
-        ----------
-        coordinates : np.ndarray, shape=(M, 3)
-            Cartesian coordinates of `M` atoms in the molecule.
-        numbers : np.ndarray, shape=(M,)
-            Atomic number of `M` atoms in the molecule.
-        pseudo_numbers : np.ndarray, shape=(M,)
-            Pseudo-number of `M` atoms in the molecule.
-        specs : str, optional
-            Specification of grid. Either choose from ['coarse', 'medium', 'fine', 'veryfine',
-            'ultrafine', 'insane'] or provide a string of 'rname:rmin:rmax:nrad:nang' format.
-            Here 'rname' denotes the type of radial grid and can be chosen from ['linear', 'exp',
-            'power'], 'rmin' and 'rmax' specify the first and last radial grid points in angstrom,
-            'nrad' specify the number of radial grid points, and 'nang' specify the number of
-            angular Lebedev-Laikov grid. The 'nang' can be chosen from (6, 14, 26, 38, 50, 74, 86,
-            110, 146, 170, 194, 230, 266, 302, 350, 434, 590, 770, 974, 1202, 1454, 1730, 2030,
-            2354, 2702, 3074, 3470, 3890, 4334, 4802, 5294, 5810).
-        k : int, optional
-            The order of the switching function in Becke's weighting scheme.
-        rotate : bool, optional
-            Whether to randomly rotate spherical grids.
-
-        """
         self._coordinates = coordinates
         self._numbers = numbers
         self._pseudo_numbers = pseudo_numbers
         self._k = k
         self._rotate = rotate
-        self.specs = specs
+        self._specs = specs
+        self._points_of_angular = points_of_angular
 
-        self._grid = BeckeMolGrid(self.coordinates, self.numbers, self.pseudo_numbers,
-                                  agspec=self.specs, k=k, random_rotate=rotate, mode='keep')
+        onedg = HortonLinear(100)
+        becke = BeckeWeights(order=self._k)
+
+        self._grid = MolGrid.horton_molgrid(self.coordinates, self.numbers,
+                                            onedg, self._points_of_angular, becke)
 
     @classmethod
     def from_molecule(cls, molecule, specs='medium', k=3, rotate=False):
-        """Initialize the class given an instance of Molecule.
-
-        Parameters
-        ----------
-        molecule : instance of Molecule
-            Instance of Molecule class.
-        specs : str, optional
-            Specification of grid. Either choose from ['coarse', 'medium', 'fine', 'veryfine',
-            'ultrafine', 'insane'] or provide a string of 'rname:rmin:rmax:nrad:nang' format.
-            Here 'rname' denotes the type of radial grid and can be chosen from ['linear', 'exp',
-            'power'], 'rmin' and 'rmax' specify the first and last radial grid points in angstrom,
-            'nrad' specify the number of radial grid points, and 'nang' specify the number of
-            angular Lebedev-Laikov grid. The 'nang' can be chosen from (6, 14, 26, 38, 50, 74, 86,
-            110, 146, 170, 194, 230, 266, 302, 350, 434, 590, 770, 974, 1202, 1454, 1730, 2030,
-            2354, 2702, 3074, 3470, 3890, 4334, 4802, 5294, 5810).
-        k : int, optional
-            The order of the switching function in Becke's weighting scheme.
-        rotate : bool, optional
-            Whether to randomly rotate spherical grids.
-
-        """
         if not isinstance(molecule, Molecule):
             raise TypeError('Argument molecule should be an instance of Molecule class.')
         coords, nums, pnums = molecule.coordinates, molecule.numbers, molecule.pseudo_numbers
@@ -101,32 +58,8 @@ class MolecularGrid(object):
 
     @classmethod
     def from_file(cls, fname, specs='medium', k=3, rotate=False):
-        """Initialize the class given an instance of Molecule.
-
-        Parameters
-        ----------
-        fname : str
-            Path to molecule's files.
-        specs : str, optional
-            Specification of grid. Either choose from ['coarse', 'medium', 'fine', 'veryfine',
-            'ultrafine', 'insane'] or provide a string of 'rname:rmin:rmax:nrad:nang' format.
-            Here 'rname' denotes the type of radial grid and can be chosen from ['linear', 'exp',
-            'power'], 'rmin' and 'rmax' specify the first and last radial grid points in angstrom,
-            'nrad' specify the number of radial grid points, and 'nang' specify the number of
-            angular Lebedev-Laikov grid. The 'nang' can be chosen from (6, 14, 26, 38, 50, 74, 86,
-            110, 146, 170, 194, 230, 266, 302, 350, 434, 590, 770, 974, 1202, 1454, 1730, 2030,
-            2354, 2702, 3074, 3470, 3890, 4334, 4802, 5294, 5810).
-        k : int, optional
-            The order of the switching function in Becke's weighting scheme.
-        rotate : bool, optional
-            Whether to randomly rotate spherical grids.
-
-        """
         mol = Molecule.from_file(fname)
         return cls.from_molecule(mol, specs, k, rotate)
-
-    def __getattr__(self, item):
-        return getattr(self._grid, item)
 
     @property
     def center(self):
@@ -163,43 +96,20 @@ class MolecularGrid(object):
         """Integration weight of grid points."""
         return self._grid.weights
 
-    def integrate(self, *value):
-        """Integrate the property evaluated on the grid points.
+    def integrate(self, *value_arrays):
+        r""" Integrate over the whole grid for given multiple value arrays.
+
+        Product of all value_arrays will be computed element-wise then
+        integrated on the grid with its weights.
+        .. math::
+            Integral = \int w(x) \prod_i f_i(x) dx
 
         Parameters
         ----------
-        value : np.ndarray
-           Property value evaluated on the grid points.
-
+        *value_arrays : np.ndarray(N, )
+            One or multiple value array to integrate.
         """
-        # temporary hack because of HORTON
-        if not isinstance(value, np.ndarray):
-            temp = value[0]
-            for item in value[1:]:
-                if item is not None:
-                    temp = temp * item
-            value = temp
-        if value.ndim != 1:
-            raise ValueError('Argument value should be a 1D array.')
-        if value.shape != (self.npoints,):
-            raise ValueError('Argument value should have ({0},) shape!'.format(self.npoints))
-        return self._grid.integrate(value)
+        return self._grid.integrate(*value_arrays)
 
     def compute_spherical_average(self, value):
-        """Compute spherical average of given value evaluated on the grid points.
-
-        Note: This method only works for atomic systems with one nuclear center.
-
-        Parameters
-        ----------
-        value : np.ndarray
-           Property value evaluated on the grid points.
-
-        """
-        if len(self.numbers) != 1:
-            raise NotImplementedError('This method only works for systems with one atom!')
-        if value.ndim != 1:
-            raise ValueError('Argument value should be a 1D array.')
-        if value.shape != (self.npoints,):
-            raise ValueError('Argument value should have ({0},) shape!'.format(self.npoints))
-        return self._grid.get_spherical_average(value)
+        raise NotImplementedError("Not yet implemented in grid")
